Overall clean up
- Rework `GLSPCommandStack`. The undo/redo methods of the command stack should never be called as the server is responsible for handling those requests. In any case we now ensure that a proper warning is logged and the model remains unchanged if the methods are called on accident
- Remove unused `IGLSPToolManager` interface
- Move glsp layouter (vbox,hbxo,freeform) into proper module to avoid unnecessary rebinding
- Drop support for stack layout ( Fixes https://github.com/eclipse-glsp/glsp/issues/444)
- Fix typos
- Ensure that everything is exported via root index
