+++ examples/workflow-standalone/app/diagram.html
@@ -1,38 +1,34 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8" />
-        <link rel="stylesheet" href="css/sprotty.css" />
-        <link rel="stylesheet" href="css/glsp-sprotty.css" />
-        <style>
-            html,
-            body {
-                width: 100%;
-                height: 100%;
-                overflow: hidden;
-            }
 
-            #sprotty {
-                position: absolute;
-                top: 0;
-                bottom: 0;
-                left: 0;
-                right: 0;
-            }
+<head>
+    <meta charset="utf-8" />
+    <style>
+        html,
+        body {
+            width: 100%;
+            height: 100%;
+            overflow: hidden;
+        }
 
-            svg {
-                width: 100%;
-                height: 100%;
-            }
-        </style>
-        <!-- support Microsoft browsers -->
-        <script src="lib/dom.js">
-            /* DOM4 */
-        </script>
-    </head>
+        #sprotty {
+            position: absolute;
+            top: 0;
+            bottom: 0;
+            left: 0;
+            right: 0;
+        }
 
-    <body>
-        <div id="sprotty"></div>
-        <script src="bundle.js"></script>
-    </body>
+        svg {
+            width: 100%;
+            height: 100%;
+        }
+    </style>
+</head>
+
+<body>
+    <div id="sprotty"></div>
+    <script src="bundle.js"></script>
+</body>
+
 </html>
