+++ packages/client/src/base/command-stack.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2022 EclipseSource and others.
+ * Copyright (c) 2019-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -13,22 +13,24 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { RedoAction, UndoAction } from '@eclipse-glsp/protocol';
-import { inject, injectable } from 'inversify';
-import { CommandStack, IActionDispatcher, SModelRoot } from 'sprotty';
-import { TYPES } from './types';
+import { injectable } from 'inversify';
+import { CommandStack, SModelRoot } from 'sprotty';
 
 @injectable()
 export class GLSPCommandStack extends CommandStack {
-    @inject(TYPES.IActionDispatcherProvider) protected actionDispatcher: () => Promise<IActionDispatcher>;
-
     override undo(): Promise<SModelRoot> {
-        this.actionDispatcher().then(dispatcher => dispatcher.dispatch(UndoAction.create()));
-        return this.thenUpdate();
+        this.logger.warn(
+            this,
+            'GLSPCommandStack.undo() was called. This should never happen as the GLSP server is responsible for handling undo requests'
+        );
+        return this.currentModel;
     }
 
     override redo(): Promise<SModelRoot> {
-        this.actionDispatcher().then(dispatcher => dispatcher.dispatch(RedoAction.create()));
-        return this.thenUpdate();
+        this.logger.warn(
+            this,
+            'GLSPCommandStack.redo() was called. This should never happen as the GLSP server is responsible for handling redo requests'
+        );
+        return this.currentModel;
     }
 }
