+++ packages/client/src/base/container-modules.ts
@@ -40,7 +40,7 @@ import modelHintsModule from '../features/hints/di.config';
 import glspHoverModule from '../features/hover/di.config';
 import layoutModule from '../features/layout/di.config';
 import glspMouseToolModule from '../features/mouse-tool/di.config';
-import { navigationModule } from '../features/navigation/di.config';
+import navigationModule from '../features/navigation/di.config';
 import glspSelectModule from '../features/select/di.config';
 import sourceModelWatcherModule from '../features/source-model-watcher/di.config';
 import toolFeedbackModule from '../features/tool-feedback/di.config';
