+++ packages/client/src/features/bounds/di.config.ts
@@ -23,12 +23,13 @@ import {
     LayoutRegistry,
     RequestBoundsCommand,
     SetBoundsCommand,
-    StackLayouter,
     VBoxLayouter
 } from 'sprotty';
-import '../../../css/command-palette.css';
 import { TYPES } from '../../base/types';
+import { FreeFormLayouter } from './freeform-layout';
 import { GLSPHiddenBoundsUpdater } from './glsp-hidden-bounds-updater';
+import { HBoxLayouterExt } from './hbox-layout';
+import { VBoxLayouterExt } from './vbox-layout';
 
 const glspBoundsModule = new ContainerModule((bind, _unbind, isBound, _rebind) => {
     configureCommand({ bind, isBound }, SetBoundsCommand);
@@ -39,9 +40,9 @@ const glspBoundsModule = new ContainerModule((bind, _unbind, isBound, _rebind) =
     bind(TYPES.Layouter).to(Layouter).inSingletonScope();
     bind(TYPES.LayoutRegistry).to(LayoutRegistry).inSingletonScope();
 
-    configureLayout({ bind, isBound }, VBoxLayouter.KIND, VBoxLayouter);
-    configureLayout({ bind, isBound }, HBoxLayouter.KIND, HBoxLayouter);
-    configureLayout({ bind, isBound }, StackLayouter.KIND, StackLayouter);
+    configureLayout({ bind, isBound }, VBoxLayouter.KIND, VBoxLayouterExt);
+    configureLayout({ bind, isBound }, HBoxLayouter.KIND, HBoxLayouterExt);
+    configureLayout({ bind, isBound }, FreeFormLayouter.KIND, FreeFormLayouter);
 });
 
 export default glspBoundsModule;
