+++ packages/client/src/features/bounds/freeform-layout.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021-2022 EclipseSource and others.
+ * Copyright (c) 2021-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -20,7 +20,7 @@ import { AbstractLayout } from 'sprotty/lib/features/bounds/abstract-layout';
 import { AbstractLayoutOptions } from 'sprotty/lib/features/bounds/layout-options';
 
 /**
- * Layouts children of a container with explit X/Y positions
+ * Layouts children of a container with explicit X/Y positions
  */
 @injectable()
 export class FreeFormLayouter extends AbstractLayout<AbstractLayoutOptions> {
