+++ packages/client/src/features/command-palette/command-palette-tool.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2021 EclipseSource and others.
+ * Copyright (c) 2019-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -25,7 +25,7 @@ export class CommandPaletteTool implements Tool {
 
     @postConstruct()
     protected postConstruct(): void {
-        this.commandPaletteKeyListener = this.createCommandPaleteKeyListener();
+        this.commandPaletteKeyListener = this.createCommandPaletteKeyListener();
     }
 
     get id(): string {
@@ -40,7 +40,7 @@ export class CommandPaletteTool implements Tool {
         this.keyTool.deregister(this.commandPaletteKeyListener);
     }
 
-    protected createCommandPaleteKeyListener(): KeyListener {
+    protected createCommandPaletteKeyListener(): KeyListener {
         return new CommandPaletteKeyListener();
     }
 }
