+++ packages/client/src/features/layout/di.config.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2019-2022 EclipseSource and others.
+ * Copyright (c) 2019-2023 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,27 +14,18 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { ContainerModule } from 'inversify';
-import { configureActionHandler, configureLayout, HBoxLayouter, TYPES, VBoxLayouter } from 'sprotty';
-import { FreeFormLayouter } from './freeform-layout';
-import { HBoxLayouterExt } from './hbox-layout';
+import { configureActionHandler } from 'sprotty';
 import {
     AlignElementsAction,
     AlignElementsActionHandler,
     ResizeElementsAction,
     ResizeElementsActionHandler
 } from './layout-elements-action';
-import { LayouterExt } from './layouter';
-import { VBoxLayouterExt } from './vbox-layout';
 
 const layoutModule = new ContainerModule((bind, _unbind, isBound, rebind) => {
     const context = { bind, isBound };
     configureActionHandler(context, ResizeElementsAction.KIND, ResizeElementsActionHandler);
     configureActionHandler(context, AlignElementsAction.KIND, AlignElementsActionHandler);
-
-    rebind(VBoxLayouter).to(VBoxLayouterExt);
-    rebind(HBoxLayouter).to(HBoxLayouterExt);
-    rebind(TYPES.Layouter).to(LayouterExt);
-    configureLayout({ bind, isBound }, FreeFormLayouter.KIND, FreeFormLayouter);
 });
 
 export default layoutModule;
