+++ packages/protocol/src/action-protocol/element-text-editing.spec.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022 STMicroelectronics and others.
+ * Copyright (c) 2022-2023 STMicroelectronics and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -104,7 +104,7 @@ describe('Element text editing actions', () => {
             it('should return an object conforming to the interface with matching properties for the given required and optional arguments', () => {
                 const expected: SetEditValidationResultAction = {
                     kind: 'setEditValidationResult',
-                    responseId: 'myRespsone',
+                    responseId: 'myResponse',
                     args: { some: 'args' },
                     status: { severity: ValidationStatus.Severity.OK }
                 };
