+++ packages/protocol/src/action-protocol/tool-palette.spec.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022 STMicroelectronics and others.
+ * Copyright (c) 2022-2023 STMicroelectronics and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -18,7 +18,7 @@ import { expect } from 'chai';
 import { TriggerEdgeCreationAction, TriggerNodeCreationAction } from './tool-palette';
 
 describe('Tool palette Actions', () => {
-    describe('TrigerNodeCreationAction', () => {
+    describe('TriggerNodeCreationAction', () => {
         describe('is', () => {
             it('should return true for an object having the correct type and a value for all required interface properties', () => {
                 const action: TriggerNodeCreationAction = {
